requirements.txt
torch>=1.13.0
torchvision>=0.14.0
streamlit>=1.20.0
opencv-python>=4.7.0.0
Pillow>=9.0.0
numpy>=1.22.0
torchcam>=0.3.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
pydicom>=2.3.0
albumentations>=1.3.0
tqdm>=4.65.0
pandas>=1.5.0
plotly>=5.10.0
transformers>=4.30.0
